Align RHIME model spec ownership
