+++ openghg_inversions/models/rhime.py
@@ -13,6 +13,7 @@ and names PyMC variables by sector.
 
 from __future__ import annotations
 
+from dataclasses import dataclass
 import re
 from collections.abc import Mapping, Sequence
 from typing import Any, cast
@@ -28,12 +29,67 @@ from openghg_inversions.models.components import (
 )
 from openghg_inversions.models.coords import CoordRegistry, attach_coord_registry
 from openghg_inversions.models.priors import PriorArgs
-from openghg_inversions.rhime.model_specs import RhimeModelSpec, SectorSpec as SectorSpec
 
 DEFAULT_X_PRIOR: PriorArgs = {"pdf": "lognormal", "mean": 1.0, "stdev": 1.0, "reparameterise": True}
 DEFAULT_BC_PRIOR: PriorArgs = {"pdf": "truncatednormal", "mu": 1.0, "sigma": 0.05, "lower": 0.0}
 DEFAULT_SIGMA_PRIOR: PriorArgs = {"pdf": "uniform", "lower": 0.1, "upper": 3.0}
 DEFAULT_OFFSET_PRIOR: PriorArgs = {"pdf": "normal", "mu": 0, "sigma": 1}
+
+
+@dataclass(frozen=True)
+class SectorSpec:
+    """Configuration for one separately optimised flux sector.
+
+    Args:
+        name: User-facing sector name.
+        flux_source: OpenGHG flux ``source`` used to retrieve this sector.
+        x_prior: Prior specification for this sector's flux scaling factors.
+        variable_suffix: PyMC-safe suffix used in multi-sector model variable
+            names. Standard single-sector RHIME uses plain ``x``/``mu`` names.
+    """
+
+    name: str
+    flux_source: str
+    x_prior: dict[str, Any]
+    variable_suffix: str
+
+
+@dataclass(frozen=True)
+class RhimeModelSpec:
+    """Model options used to build a RHIME PyMC model.
+
+    Args:
+        species: Primary gas or tracer name used for object-store lookup and
+            output naming.
+        domain: Model domain name.
+        sectors: Flux sectors included in the model. Each sector is optimized
+            separately and is normally backed by one OpenGHG flux ``source``.
+        use_bc: Whether boundary-condition scaling is included.
+        sigma_per_site: Whether model-error terms vary by site.
+        add_offset: Whether model-data offsets are included.
+        pollution_events_from_obs: Whether model error scales with observed
+            enhancements instead of modelled enhancements.
+        no_model_error: Whether explicit model-error terms are disabled.
+        power: Exponent or prior specification used in likelihood error scaling.
+        bc_prior: Prior specification for boundary-condition scaling factors.
+        sigma_prior: Prior specification for model-error terms.
+        offset_prior: Prior specification for optional offsets.
+        offset_args: Extra keyword arguments forwarded to the offset component.
+    """
+
+    species: str
+    domain: str
+    sectors: tuple[SectorSpec, ...]
+    use_bc: bool = True
+    sigma_per_site: bool = True
+    add_offset: bool = False
+    pollution_events_from_obs: bool = False
+    no_model_error: bool = False
+    power: dict[str, Any] | float = 1.99
+    bc_prior: dict[str, Any] | None = None
+    sigma_prior: dict[str, Any] | None = None
+    offset_prior: dict[str, Any] | None = None
+    offset_args: dict[str, Any] | None = None
 
 
 def safe_pymc_name(value: str) -> str:
